rename `git_client` to `git`
